+++ Editor/UI/AssetManagerWindow.cs
@@ -10,7 +10,6 @@ namespace Unity.AssetManager.Editor
 
         public static AssetManagerWindow instance { get; private set; }
         private AssetManagerWindowRoot m_Root;
-        private ILinksProxy m_LinksProxy;
 
         [MenuItem("Window/Asset Manager", priority = 1500)]
         internal static void Open()
@@ -47,7 +46,6 @@ namespace Unity.AssetManager.Editor
                 container.Resolve<IAssetDatabaseProxy>());
             m_Root.OnEnable();
             m_Root.StretchToParentSize();
-            m_LinksProxy = container.Resolve<ILinksProxy>();
             rootVisualElement.Add(m_Root);
         }
 
@@ -88,12 +86,7 @@ namespace Unity.AssetManager.Editor
 
         public void AddItemsToMenu(GenericMenu menu)
         {
-            GUIContent goToDashboard = new GUIContent(L10n.Tr("Go to Dashboard"));
-            menu.AddItem(goToDashboard, false, m_LinksProxy.OpenAssetManagerDashboard);
-            GUIContent projectSettings = new GUIContent(L10n.Tr("Project Settings"));
-            menu.AddItem(projectSettings, false, m_LinksProxy.OpenProjectSettingsServices);
-            GUIContent preferences = new GUIContent(L10n.Tr("Preferences"));
-            menu.AddItem(preferences, false, m_LinksProxy.OpenPreferences);
+            m_Root?.AddItemsToMenu(menu);
         }
     }
 }